Fix NT_GNU_BUILD_ID check
The ELF note containing the build id is currently identified only based
on the n_type field. However, the interpretation of the type depends on
the namespace defined by the name field; types are not globally unique.

For example, ".note.hwasan.globals" and ".note.gnu.build-id have" have
the same type:

  NT_LLVM_HWASAN_GLOBALS = 3
  NT_GNU_BUILD_ID = 3

In a binary compiled with -fsanitize=hwaddress, breakpad therefore reads
the build id from the wrong section.

Fix this by matching on the name ("GNU") in addition to type. This is
consistent with both LLVM[1] and libunwindstack[2].

[1] https://github.com/llvm/llvm-project/blob/823eb1a3/llvm/lib/Object/BuildID.cpp#L38
[2] https://android.googlesource.com/platform/system/unwinding/+/refs/tags/android-14.0.0_r33/libunwindstack/ElfInterface.cpp#252

Test: dumper_unittest
Change-Id: I2fed025f07efcd5699324695de2e8b65331c9707