Adds a Flag to Perserve Address Offset in dump_syms
By default, the .code section start address is normalized to zero,
and subsequently the address of all functions and lines are adjusted
accordingly. This change adds a command line flag that perserves the
original addresses. This is a requirement for those that may wish to use
the crash resolver, but do not have a minidump file to work with (most commonly in firmware).

TEST=make check
BUG=b:328511413

Change-Id: Id7c477196ffed7fd319029b28ed2607192249c6c