Adds Support for EC Zephyr Binaries
The text section header for EC Zephyr binaries is `text` instead of the
standard `.text`. As a result the file_id hash function was failing when
attempting to generate a file id by hashing the contents of a Zephyr
binaries `text` section.

This change adds support for Zephyr binaries by having said hash
function also check the `text` section in addition to `.text`.

TEST=make check && dump_syms -a ~/some.zephyr.elf
BUG=b:328511413

Change-Id: I88882305a4c9ca18082ba2690ba1b22da03bd19d