Make symupload.exe build so it can be used with current versions of the Visual Studio msdia140.dll. We don't use the V2 uploader, so I removed it because it was causing a linker error.
This commit includes all of the Visual Studio build files generated
by the batchfile in the tools/windows directory. I'm doing this because
the tool that generates them is no longer maintained and isn't the
easiest thing to find documentation for.

This branch also has a reverted commit - the upstream maintainers of
breakpad removed all of the .gyp files, seemingly without noticing
that doing so broke the build script for these tools.

Hopefully nobody ever has to build this again, but if they do I hope
this is enough information to steer them in the right direction.